Fix three streaming engine bugs: key mismatch, O(N^2) cipher, uint8 overflow
StreamServer/Client produced different keys because each side only
accumulated its own Z noise samples. Extended wire frames to carry
encrypted z_k values so both parties accumulate (z_bob, z_alice) in
the same interleaved order.

Replaced AuthCipher's O(N^2) SHAKE-256 digest (re-generating from
byte 0 on every call) with counter-mode chunked generation: O(1) per
8KB chunk.

Fixed uint8 overflow in Toeplitz matrix-vector multiply when
n_raw > 255 by using uint16 storage for the hash matrix.

+++ src/liuproto/privacy.py
@@ -53,7 +53,7 @@ class PrivacyAmplification:
         if precompute:
             self._matrix = np.array(
                 [self._toeplitz_row(i) for i in range(n_secure)],
-                dtype=np.uint8)
+                dtype=np.uint16)
 
     def _toeplitz_row(self, i):
         """Return row i of the Toeplitz matrix (length n_raw)."""
@@ -82,7 +82,7 @@ class PrivacyAmplification:
             raise ValueError("Expected %d raw bits, got %d" %
                              (self.n_raw, len(raw)))
         if self._matrix is not None:
-            return (self._matrix @ raw) % 2
+            return ((self._matrix @ raw) % 2).astype(np.uint8)
         return self._hash_fft(raw)
 
     def hash_fast(self, raw_bits):
@@ -112,10 +112,10 @@ class PrivacyAmplification:
             if mem_bytes <= 100_000_000:  # ~100 MB threshold
                 self._matrix = np.array(
                     [self._toeplitz_row(i) for i in range(self.n_secure)],
-                    dtype=np.uint8)
-                return (self._matrix @ raw) % 2
+                    dtype=np.uint16)
+                return ((self._matrix @ raw) % 2).astype(np.uint8)
             return self._hash_fft(raw)
-        return (self._matrix @ raw) % 2
+        return ((self._matrix @ raw) % 2).astype(np.uint8)
 
     def _hash_fft(self, raw):
         """Toeplitz hash via FFT-based convolution in O(n log n).
